Created ActionButtons and Couples component and use reactstrap select
