+++ tennisblock_client/src/Schedule/modules/schedule/selectors.js
@@ -6,48 +6,14 @@ export const getCouples = ({ blockplayers }) => ({couples: [...blockplayers.coup
 
 export const getSubs = ({ subs }) => ({subs: [...subs.guysubs, ...subs.galsubs] })
 
+export const getGuySubs = ({ subs_guys }) => (subs_guys)
+
+export const getGalSubs = ({ subs_gals }) => (subs_gals)
+
 export const isBlockPlayerEdited = ({ blockplayers, originalCouples }) => {
   const couples = Object.assign({}, blockplayers.couples)
   const base = Object.assign({}, originalCouples)
   return JSON.stringify(base) !== JSON.stringify(couples)
-}
-
-export const getGuySubOptions = ({ subs, blockplayers }) => {
-  const { couples } = blockplayers
-  const { guysubs } = subs
-
-  return (couples || []).reduce((result, value, index) => {
-    const guyPlayers = guysubs && [couples[index].guy].concat(guysubs)
-    result[index] = guyPlayers && guyPlayers.map((player) => {
-      return {
-        label: player.name,
-        value: player.id,
-        key: Number(index),
-        gender: 'guy',
-        player
-      }
-    })
-    return result
-  }, [])
-}
-
-export const getGalSubOptions = ({ subs, blockplayers }) => {
-  const { couples } = blockplayers
-  const { galsubs } = subs
-
-  return (couples || []).reduce((result, value, index) => {
-    const galPlayers = galsubs && [couples[index].gal].concat(galsubs)
-    result[index] = galPlayers && galPlayers.map((player) => {
-      return {
-        label: player.name,
-        value: player.id,
-        key: Number(index),
-        gender: 'gal',
-        player
-      }
-    })
-    return result
-  }, [])
 }
 
 export const changes = ({ originalCouples, blockplayers }) => {
