+++ tennisblock_client/src/containers/MeetingSchedule/styles.local.scss
@@ -1,3 +1,7 @@
 .tableHeader {
   font-size: 16px;
 }
+
+.changed {
+  background: #E8E8E8;
+}
